+++ Cargo.toml
@@ -36,3 +36,4 @@ whoami = ">=0.9.0"
 string_morph = ">=0.1.0"
 walkdir = ">=2.3.1"
 convert_case = ">=0.4.0"
+json-structural-diff = ">=0.1.0"
